0401 mypage modify

+++ src/main/java/com/health/keeper/repository/UserRepository.java
@@ -3,7 +3,7 @@ package com.health.keeper.repository;
 import com.health.keeper.entity.UserEntity;
 import org.springframework.data.jpa.repository.JpaRepository;
 
-public interface UserRepository extends JpaRepository<UserEntity, Integer> { //<엔티티이름, pk 타입>
+public interface UserRepository extends JpaRepository<UserEntity, Long> { //<엔티티이름, pk 타입>
 
     public UserEntity findByUsername(String username);
